Init event listeners etter CSR sjekk
