+++ packages/client/src/window-event-handlers.ts
@@ -1,0 +1,91 @@
+import { env, param, updateDecoratorParams } from "./params";
+import { formatParams } from "decorator-shared/json";
+import type { Context, ParamKey, Params } from "decorator-shared/params";
+import { createEvent } from "./events";
+import { useLoadIfActiveSession } from "./screensharing";
+import Cookies from "js-cookie";
+import { addFaroMetaData } from "./faro";
+import { fetchAndRenderClientSide } from "./csr";
+
+const shouldReRender = ({ language, context }: Partial<Params>) =>
+    (language && language !== param("language")) ||
+    (context && context !== param("context"));
+
+const msgSafetyCheck = (message: MessageEvent) => {
+    const { origin, source } = message;
+    // Only allow messages from own window
+    return window.location.href.startsWith(origin) && source === window;
+};
+
+export const initWindowEventHandlers = () => {
+    window.addEventListener("paramsupdated", (e) => {
+        if (shouldReRender(e.detail.params)) {
+            fetchAndRenderClientSide(
+                `${env("APP_URL")}/env?${formatParams(window.__DECORATOR_DATA__.params)}`,
+            );
+        }
+    });
+
+    window.addEventListener("message", (e) => {
+        if (!msgSafetyCheck(e)) {
+            return;
+        }
+        if (e.data.source === "decoratorClient" && e.data.event === "ready") {
+            window.postMessage({ source: "decorator", event: "ready" });
+        }
+        if (e.data.source === "decoratorClient" && e.data.event == "params") {
+            const payload = e.data.payload;
+
+            (
+                [
+                    "breadcrumbs",
+                    "availableLanguages",
+                    "utilsBackground",
+                    "language",
+                    "chatbotVisible",
+                ] satisfies ParamKey[]
+            ).forEach((key) => {
+                if (payload[key] !== undefined) {
+                    updateDecoratorParams({
+                        [key]: payload[key],
+                    });
+                }
+            });
+
+            if (e.data.payload.context) {
+                const context = e.data.payload.context;
+                if (
+                    [
+                        "privatperson",
+                        "arbeidsgiver",
+                        "samarbeidspartner",
+                    ].includes(context)
+                ) {
+                    window.dispatchEvent(
+                        createEvent("activecontext", {
+                            bubbles: true,
+                            detail: { context },
+                        }),
+                    );
+                } else {
+                    console.warn("Unrecognized context", context);
+                }
+            }
+        }
+    });
+
+    window.addEventListener("activecontext", (event) => {
+        updateDecoratorParams({
+            context: (event as CustomEvent<{ context: Context }>).detail
+                .context,
+        });
+    });
+
+    // @TODO: Refactor loaders
+    window.addEventListener("load", () => {
+        useLoadIfActiveSession({
+            userState: Cookies.get("psCurrentState"),
+        });
+        addFaroMetaData();
+    });
+};
